pipelines/requirements.txt
ipython>=8.10
jupyterlab>=3.0
kedro~=0.19.12
kedro-viz>=6.7.0
kedro-datasets[pandas-csvdataset, pandas-exceldataset, pandas-parquetdataset]>=3.0; python_version >= "3.9"
kedro-datasets[pandas.CSVDataset, pandas.ExcelDataset, pandas.ParquetDataset, json.JSONDataset]>=1.0; python_version < "3.9"
notebook
google-cloud-firestore==2.27.0
google-cloud-storage==2.18.0
scikit-learn~=1.5.1
numpy==1.22.4
pandas-gbq==0.23.2
gcsfs==2024.9.0.post1
jinja2==3.1.3
lightgbm==3.3.2